+++ src/utils/markdown/extractAllBlocksFromMarkdown.ts
@@ -7,6 +7,11 @@ import { capitalize } from '../normalization/capitalize';
  */
 export type CodeBlock = {
     /**
+     * Which notation was used to open the code block
+     */
+    readonly blockNotation: '```' | '>';
+
+    /**
      * Language of the code block OR null if the language is not specified in opening ```
      */
     readonly language: string | null;
@@ -33,14 +38,36 @@ export function extractAllBlocksFromMarkdown(markdown: string_markdown): Array<C
     const codeBlocks: Array<CodeBlock> = [];
     const lines = markdown.split('\n');
 
+    // Note: [0] Ensure that the last block notated by gt > will be closed
+    lines.push('');
+
     let currentCodeBlock: Writable<CodeBlock> | null = null;
 
     for (const line of lines) {
+        if (line.startsWith('> ')) {
+            if (currentCodeBlock === null) {
+                currentCodeBlock = { blockNotation: '>', language: null, content: '' };
+            } /* not else */
+
+            if (currentCodeBlock.blockNotation === '>') {
+                if (currentCodeBlock.content !== '') {
+                    currentCodeBlock.content += '\n';
+                }
+
+                currentCodeBlock.content += line.slice(2);
+            }
+        } else if (currentCodeBlock !== null && currentCodeBlock.blockNotation === '>' /* <- Note: [0] */) {
+            codeBlocks.push(currentCodeBlock);
+            currentCodeBlock = null;
+        }
+
+        /* not else */
+
         if (line.startsWith('```')) {
             const language = line.slice(3).trim() || null;
 
             if (currentCodeBlock === null) {
-                currentCodeBlock = { language, content: '' };
+                currentCodeBlock = { blockNotation: '```', language, content: '' };
             } else {
                 if (language !== null) {
                     // [🌻]
@@ -53,7 +80,7 @@ export function extractAllBlocksFromMarkdown(markdown: string_markdown): Array<C
                 codeBlocks.push(currentCodeBlock);
                 currentCodeBlock = null;
             }
-        } else if (currentCodeBlock !== null) {
+        } else if (currentCodeBlock !== null && currentCodeBlock.blockNotation === '```') {
             if (currentCodeBlock.content !== '') {
                 currentCodeBlock.content += '\n';
             }
@@ -71,3 +98,7 @@ export function extractAllBlocksFromMarkdown(markdown: string_markdown): Array<C
 
     return codeBlocks;
 }
+
+/**
+ * TODO: Maybe name for `blockNotation` instead of  '```' and '>'
+ */